requirements.txt
asgiref==3.7.2
Django==5.0.2
django-debug-toolbar==4.3.0
django-filter==23.5
django-js-asset==2.2.0
django-mathfilters==1.0.0
django-mptt==0.16.0
django-phonenumber-field==7.3.0
flake8==7.0.0
mccabe==0.7.0
mysql-connector-python==8.3.0
mysqlclient==2.2.4
phonenumbers==8.13.32
pillow==10.2.0
pycodestyle==2.11.1
pyflakes==3.2.0
python-dotenv==1.0.1
pytz==2023.3.post1
sqlparse==0.4.4
tzdata==2023.4
